Added partitioning guide
According to the current version of the draft RHEL7 STIG, there is still going to be a requirement for CCI-000336 (even though the OpenSCAP guide seems to think otherwise, I have been dinged in on it before in an assessment). This was also a present control in the RHEL5 and RHEL6 STIGs. Although, in the RHEL5 STIG it has been identified as CCI-001208.

I will test this change to the kickstart before submitting a pull request.